v1.2.3: Python API refactoring, CRAM bug fix, improved error handling
Major changes:
- Fix CRAM field order bug that caused reads to be incorrectly filtered
- Refactor Python package for consistent naming (normalize→compute_reactivity,
  all→plot_all, visualize→visualize_structure)
- Add ProbingData.load() for loading from HDF5 files
- Add working cmuts-plot and cmuts-visualize CLI commands
- Add Alignment::create() factory method to prevent initialization bugs
- Add structured logging with DEBUG/INFO/WARN/ERROR levels
- Add validation macros for bounds checking
- Replace legacy bash tests with pytest integration tests
- Add Linux compatibility fixes
- Update documentation with Linux instructions and test guide

+++ src/cpp/fasta.hpp
@@ -39,7 +39,7 @@ const size_t FASTA_DATASET_SIZE = 2;
 struct Offset {
 
     int32_t length = 0;
-    int32_t offset = 0;
+    int64_t offset = 0;
 
 };
 
@@ -61,7 +61,7 @@ class Header{
 private:
 
     std::vector<HeaderBlock> _blocks;
-    int32_t _offset = 0;
+    int64_t _offset = 0;
 
 public:
